Restructured the app absolute panels
I realized that the structure of the app was making me run the rose plot and umap a number of unnecessary times. To fix that, I restructured the app so that the server side for those plots is only run once. I also added the density to plot to the filtering tool. Now the download button might not work and I will have to fix it accordingly. Also, reducing the memory usage might be useful too.