Fixing issue with wrong object input.
The code was always attempting to turn the input object into a data.frame, and this works most of the time (even when the object is something absurd like an integer). Now the code checks the input object and if it is not valid, it allows you to still load an object.